refactor(gnrwebstruct): split monolithic module into package
Promote `gnrpy/gnr/web/gnrwebstruct.py` (3186 lines) to package
`gnrpy/gnr/web/gnrwebstruct/`, on the model of `gnrpy/gnr/sql/gnrsql/`.

Package layout:
- __init__.py        — re-exports all public symbols
- _helpers.py        — cellFromField, _selected_defaultFrom, externalStore
- base.py            — GnrDomSrc + StructMethodError + struct_method
                       + GnrDomSrcError + GnrDomElem
- dojo11.py          — GnrDomSrc_dojo_11
- dojo20.py          — GnrDomSrc_dojo_20 stub (kept separate to mark the
                       d11/d20 alignment boundary)
- formbuilder.py     — GnrFormBuilder
- gridstruct.py      — GnrGridStruct

The new package is registered in gnrpy/pyproject.toml under
[tool.setuptools].packages, mirroring how gnr.sql.gnrsql was registered
when the SQL monolith was split.

Public API is preserved: `from gnr.web.gnrwebstruct import X` continues
to work for every previously exported symbol (verified by the
characterization test suite added in the previous commit, all 17 pass).

Behavioural diff vs the old monolith: zero. genroNameSpace contains 256
entries before and after; __getattr__ resolves every widget identically.

The trailing `if __name__ == '__main__':` demo block (lines 3170-3186 of
the old file) is dropped — it referenced a hard-coded path
`/usr/local/genro/data/instances/assopy` that has not existed for years,
and a package directory cannot be executed as a script anyway.

base.py is intentionally kept at ~1100 lines: PR #3 will trim it by
replacing trivial HTML widget methods (h1-h6, div, span, ...) with
namespace-driven dispatch via the @widget decorator.

Refs #901
