+++ gnrpy/gnr/web/gnrwebstruct/__init__.py
@@ -1,0 +1,50 @@
+#-*- coding: utf-8 -*-
+
+#--------------------------------------------------------------------------
+# package       : GenroPy web - see LICENSE for details
+# module        : Genro Web structures - public package surface
+# Copyright (c) : 2004 - 2007 Softwell sas - Milano
+# Written by    : Giovanni Porcari, Michele Bertoldi
+#                 Saverio Porcari, Francesco Porcari , Francesco Cavazzana
+#--------------------------------------------------------------------------
+
+#This library is free software; you can redistribute it and/or
+#modify it under the terms of the GNU Lesser General Public
+#License as published by the Free Software Foundation; either
+#version 2.1 of the License, or (at your option) any later version.
+
+#This library is distributed in the hope that it will be useful,
+#but WITHOUT ANY WARRANTY; without even the implied warranty of
+#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+#Lesser General Public License for more details.
+
+#You should have received a copy of the GNU Lesser General Public
+#License along with this library; if not, write to the Free Software
+#Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
+
+from gnr.web.gnrwebstruct._helpers import cellFromField
+from gnr.web.gnrwebstruct.base import (
+    GnrDomElem,
+    GnrDomSrc,
+    GnrDomSrcError,
+    StructMethodError,
+    struct_method,
+)
+from gnr.web.gnrwebstruct.dojo11 import GnrDomSrc_dojo_11
+from gnr.web.gnrwebstruct.dojo20 import GnrDomSrc_dojo_20
+from gnr.web.gnrwebstruct.formbuilder import GnrFormBuilder
+from gnr.web.gnrwebstruct.gridstruct import GnrGridStruct
+
+
+__all__ = [
+    'GnrDomElem',
+    'GnrDomSrc',
+    'GnrDomSrc_dojo_11',
+    'GnrDomSrc_dojo_20',
+    'GnrDomSrcError',
+    'GnrFormBuilder',
+    'GnrGridStruct',
+    'StructMethodError',
+    'cellFromField',
+    'struct_method',
+]
