+++ gnrpy/gnr/web/gnrwebstruct/dojo20.py
@@ -1,0 +1,13 @@
+#-*- coding: utf-8 -*-
+
+#--------------------------------------------------------------------------
+# package       : GenroPy web - see LICENSE for details
+# module        : Genro Web structures - GnrDomSrc_dojo_20 stub
+# Copyright (c) : 2004 - 2007 Softwell sas - Milano
+#--------------------------------------------------------------------------
+
+from gnr.web.gnrwebstruct.dojo11 import GnrDomSrc_dojo_11
+
+
+class GnrDomSrc_dojo_20(GnrDomSrc_dojo_11):
+    pass
