+++ gnrpy/pyproject.toml
@@ -156,7 +156,8 @@ packages = [
 	    'gnr.web.gnrwebpage_plugin',
 	    'gnr.web.gnrwebpage_proxy',
 	    'gnr.web.gnrwebpage_proxy.apphandler',
-    	    'gnr.web.gnrwebpage_proxy.frontend',	
+    	    'gnr.web.gnrwebpage_proxy.frontend',
+	    'gnr.web.gnrwebstruct',
 	    'gnr.web.gnrwsgisite_proxy',
 	    'gnr.web.widgets',
 	    'gnr.xtnd',
